Feature/folder structuring (#19)
* Initial commit

* Restructuring the contracts folder

* Restructuring the contracts folder after compilation check

+++ contracts/models/Schema.sol
@@ -18,42 +18,37 @@ pragma solidity 0.7.0;
 */
 
 struct Round {
-   uint Id;
-   uint CompanyId;
-   uint LockUpPeriodForShare;
-   uint256[] PricePerShare;
-   address[] PaymentCurrencies;
-   uint TotalTokensUpForSale;
-   uint TotalInvestors;
-   uint[] TotalRaised;
-   uint TotalTokensSold;
-   uint RoundStartTimeStamp;   
-   uint DurationInSeconds;
-   string DocumentUrl;
-
-   // When set to true, the round duration is not considered
-   // The round is kept open until it has been fully subscribed
-   // Fully subscribed being that the TotalTokensUpForSale == TotalRaised
-   bool RunTillFullySubscribed;
-   bool IsDeleted;
+    uint256 Id;
+    uint256 CompanyId;
+    uint256 LockUpPeriodForShare;
+    uint256[] PricePerShare;
+    address[] PaymentCurrencies;
+    uint256 TotalTokensUpForSale;
+    uint256 TotalInvestors;
+    uint256[] TotalRaised;
+    uint256 TotalTokensSold;
+    uint256 RoundStartTimeStamp;
+    uint256 DurationInSeconds;
+    string DocumentUrl;
+    // When set to true, the round duration is not considered
+    // The round is kept open until it has been fully subscribed
+    // Fully subscribed being that the TotalTokensUpForSale == TotalRaised
+    bool RunTillFullySubscribed;
+    bool IsDeleted;
 }
 
-struct RoundInvestment
-{
-   uint256 RoundId;
-   uint256 TokenAlloaction;
-   address[] PaymentCurrencies;
-   uint256[] InvestmentAmounts;
-   bool Exists;
+struct RoundInvestment {
+    uint256 RoundId;
+    uint256 TokenAlloaction;
+    address[] PaymentCurrencies;
+    uint256[] InvestmentAmounts;
+    bool Exists;
 }
 
-
-struct Index
-{
-   uint256 Index;
-   bool Exists;
+struct Index {
+    uint256 Index;
+    bool Exists;
 }
-
 
 /**
 
@@ -63,102 +58,84 @@ struct Index
   Lots of integration potentials here with other nft market places
 
  */
-struct RoundNft
-{
-   uint Id;
-   uint RoundId;
-   address NftContractAddress;
-   address Investor;
-
+struct RoundNft {
+    uint256 Id;
+    uint256 RoundId;
+    address NftContractAddress;
+    address Investor;
 }
 
-struct Proposal{
-   uint Id;
-   uint CompanyId;
-   uint[] AmountRequested;
-   address[] PaymentCurrencies;
-   uint VoteSessionDuration;
-   uint VoteStartTimeStamp;
-   uint ApprovedVotes;
-   uint RejectedVotes;
-   uint TokensStakedForApprovedVotes;
-   uint TokensStakedForRejectedVotes;
-   bool IsDeleted;
-   bool HasWithdrawn;
-
+struct Proposal {
+    uint256 Id;
+    uint256 CompanyId;
+    uint256[] AmountRequested;
+    address[] PaymentCurrencies;
+    uint256 VoteSessionDuration;
+    uint256 VoteStartTimeStamp;
+    uint256 ApprovedVotes;
+    uint256 RejectedVotes;
+    uint256 TokensStakedForApprovedVotes;
+    uint256 TokensStakedForRejectedVotes;
+    bool IsDeleted;
+    bool HasWithdrawn;
 }
 
-struct Investor{
-   address WalletAddress;
-   uint256 proposalsApproved;
-   uint256 proposalsRejected;
-
+struct Investor {
+    address WalletAddress;
+    uint256 proposalsApproved;
+    uint256 proposalsRejected;
 }
 
-struct Company{
-   uint Id;
-   string CompanyName;
-   string CompanyUrl;
-   address CompanyTokenContractAddress;
-   address OwnerAddress;
+struct Company {
+    uint256 Id;
+    string CompanyName;
+    string CompanyUrl;
+    address CompanyTokenContractAddress;
+    address OwnerAddress;
 }
 
-struct SupportedPaymentOption
-{
-   bool IsEnabled;
-   bool Exists;
-   uint256 Index;
+struct SupportedPaymentOption {
+    bool IsEnabled;
+    bool Exists;
+    uint256 Index;
 }
 
-struct ProposalVote
-{
-   uint256 ProposalId;  
-   uint256 SharesStaked;
-   bool IsApproved;
-   bool Exists;
+struct ProposalVote {
+    uint256 ProposalId;
+    uint256 SharesStaked;
+    bool IsApproved;
+    bool Exists;
 }
-
-
-
 
 // Service Models
 
-struct ProposalResponse
-{
-   uint ApprovedVotes;
-   uint RejectedVotes;
-   uint TokensStakedForApprovedVotes;
-   uint TokensStakedForRejectedVotes;
-   bool IsProposalApproved;
-   bool HasVotingPeriodElapsed;
-
+struct ProposalResponse {
+    uint256 ApprovedVotes;
+    uint256 RejectedVotes;
+    uint256 TokensStakedForApprovedVotes;
+    uint256 TokensStakedForRejectedVotes;
+    bool IsProposalApproved;
+    bool HasVotingPeriodElapsed;
 }
 
-struct RoundResponse
-{
-   uint Id;
-   uint CompanyId;
-   uint LockUpPeriodForShare;
-   uint256[] PricePerShare;
-   address[] PaymentCurrencies;
-
-   uint TotalTokensUpForSale;
-   uint TotalInvestors;
-   uint[] TotalRaised;
-   uint TotalTokensSold;
-   uint RoundStartTimeStamp;   
-   uint DurationInSeconds;
-   string DocumentUrl;
-   bool RunTillFullySubscribed;
-   bool IsOpen;
+struct RoundResponse {
+    uint256 Id;
+    uint256 CompanyId;
+    uint256 LockUpPeriodForShare;
+    uint256[] PricePerShare;
+    address[] PaymentCurrencies;
+    uint256 TotalTokensUpForSale;
+    uint256 TotalInvestors;
+    uint256[] TotalRaised;
+    uint256 TotalTokensSold;
+    uint256 RoundStartTimeStamp;
+    uint256 DurationInSeconds;
+    string DocumentUrl;
+    bool RunTillFullySubscribed;
+    bool IsOpen;
 }
 
-struct RebalancedProposalPayout
-{
-   address currencyAddress;
-   uint256 amountToSend;
+struct RebalancedProposalPayout {
+    address currencyAddress;
+    uint256 amountToSend;
 }
-
-
-
-
